Fix question index consistency and score/stat clamping

+++ app/src/main/java/com/example/ledgerscanner/feature/scanner/scan/viewmodel/OmrScannerViewModel.kt
@@ -318,7 +318,7 @@ class OmrScannerViewModel @Inject constructor(
                     scannedAt = System.currentTimeMillis(),
                     studentAnswers = studentAnswers,
                     multipleMarksDetected = evaluation.multipleMarksQuestions,
-                    score = evaluation.marksObtained.toInt(),
+                    score = evaluation.marksObtained,
                     totalQuestions = evaluation.totalQuestions,
                     correctCount = evaluation.correctCount,
                     wrongCount = evaluation.incorrectCount,
